requirements.txt
click
Jinja2
pyaml